requirements/base.txt
alembic==1.13.1
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
python-decouple==3.8
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
